Fixed bug with balance

+++ models/wallet.go
@@ -1,9 +1,9 @@
 package models
 
 type Wallet struct {
-	ID       int64  `json:"id" field:"id"`
-	OwnerID  int64  `json:"owner_id" field:"ownerid"`
-	Name     string `json:"name" validate:"omitempty,max=50" field:"name"`
-	Currency string `json:"currency" validate:"omitempty,max=50" field:"currency"`
-	Balance  int64  `json:"balance" field:"balance"`
+	ID       int64   `json:"id" field:"id"`
+	OwnerID  int64   `json:"owner_id" field:"ownerid"`
+	Name     string  `json:"name" validate:"omitempty,max=50" field:"name"`
+	Currency string  `json:"currency" validate:"omitempty,max=50" field:"currency"`
+	Balance  float64 `json:"balance" field:"balance"`
 }
